Fixed missing ItemTableThink
Earlier changes to attribute applications made custom attributes not work for custom weapons, since there was no ItemTableThink applied to them. This adds a few lines of code to address this, so an AddThinkToEnt is slapped to the item after it's DispatchSpawned.
Also comments out the ExtraItems listed within so you can use ExtraItems as intended.